Cache the size of part_list/scavd_list (#11783)
After a parallel GC, it is possible to have a long list of blocks in
ws->part_list, if we did a lot of work stealing but didn't fill up the
blocks we stole.  These blocks persist until the next load-balanced GC,
which might be a long time, and during every GC we were traversing this
list to find its size.  The fix is to maintain the size all the time, so
we don't have to compute it.

(cherry picked from commit 5c4cd0e44657d52f7ca5fee63f8765d17f1fbe85)