Use tcExtendGlobalValEnv for default methods
We add the default method Ids to the global type envt in
tcAddImplicits; and then again in tcInstDcls2 (for reasons
explained there).  But for some reason in the latter case
we added them to the /local/ type envt, via tcExtendLetEnv.

This patch just uses tcExtendGlobalValEnv consistently
for both.

(I found this when reviewing slightly-awkward code from Facundo's
static-form error-message patch; turned out that the awkwardness
came back to this inconsistency in the handling of default
methods.)